feat: refactor message handling to support image uploads with message IDs
